vkd3d: Use vkd3d_atomic_increment_u32() in vkd3d_bind_heap_memory().

+++ libs/vkd3d/vkd3d_private.h
@@ -666,7 +666,7 @@ struct d3d12_heap
 {
     ID3D12Heap ID3D12Heap_iface;
     unsigned int refcount;
-    LONG resource_count;
+    unsigned int resource_count;
 
     bool is_private;
     D3D12_HEAP_DESC desc;
